Support GAB LDAP Stored certificates
This commit adds support for fetching public certificates from LDAP for
GAB users for encrypting and verification. When encrypting or verifying
a message the LDAP certificate is first looked up, before the user store
certificates.

When a GAB certificate is avaliable, we will not import the public
certificate when viewing a signed email or uploading a private
certificate.

References: KSP-8