php: Use EncryptionStore for passphrase caching
Cache the passphrase for 5 minutes in the EncryptionStore so that the
user has the time to download multiple attachments without having to
enter their passphrase again. Note that this does still expose a bug
where after the 5 minute period the attachment can't be downloaded and
an error will be shown.

References: KSP-155