php: handle enveloped-data smime-type's
According to https://tools.ietf.org/html/rfc8551#section-3.5 signed-only
mails can also be in an envelope, as an alternative to multipart/signed.
However Core marks these as IPM.Note which makes the smime plugin think
it's encrypted while it's not. Handle this by checking for
enveloped-data and then verifying the message.

References: KSP-193